refactored bot code
