+++ package.json
@@ -22,5 +22,6 @@
     "discord.js": "^12.4.1",
     "ffmpeg-static": "^4.2.7",
     "shelljs": "^0.8.4"
-  }
+  },
+  "devDependencies": {}
 }
